Find spatially variable genes

Identifies features that have spatially variation along spots.

FindSVGs(seu, nfeatures=2000, covariates=NULL, num_core=1, verbose=TRUE)

Arguments

seu
an object of class "Seurat".

nfeatures
a positive integer, means how many spatially variable genes to be chosen. If there are less than 2000 features in seu, then all features are identified.

covariates
a covariate matrix named control variable matrix whose number of rows is equal to the number of columns of seu.

num_core
an optional positive integer, specify the cores used for identifying the SVGs in parallel.

verbose
an optional logical value, whether output the related information.

Details

Nothing

Value

return a revised Seurat object by adding three columns named "is.SVGs", "order.SVGs" and "adjusted.pval.SVGs" in the meta.features of default Assay.

References

Zhu, J., Sun, S., Zhou, X.: Spark-x: non-parametric modeling enables scalable and robust detection of spatialexpression patterns for large spatial transcriptomic studies. Genome Biology 22(1), 1–25 (2021)

Note

nothing

See also

Examples

  seu <- gendata_RNAExp(height=20, width=20,p=200, K=4)
  
  seu <- FindSVGs(seu, nfeatures=100)

#> Find the spatially variables genes by SPARK-X...

#> ## ===== SPARK-X INPUT INFORMATION ====

#> ## number of total samples: 400

#> ## number of total genes: 200

#> ## Running with single core, may take some time 

#> ## Testing With Projection Kernel

#> ## Testing With Gaussian Kernel 1

#> ## Testing With Gaussian Kernel 2

#> ## Testing With Gaussian Kernel 3

#> ## Testing With Gaussian Kernel 4

#> ## Testing With Gaussian Kernel 5

#> ## Testing With Cosine Kernel 1

#> ## Testing With Cosine Kernel 2

#> ## Testing With Cosine Kernel 3

#> ## Testing With Cosine Kernel 4

#> ## Testing With Cosine Kernel 5

  topSVGs(seu)

#> [1] "gene91"  "gene99"  "gene31"  "gene61"  "gene106"