Joint dimension reduction and spatial clustering

Joint dimension reduction and spatial clustering for scRNA-seq and spatial transcriptomics data

DR.SC(object, ...)
### S3 method for class "Seurat"
DR.SC(seu, q=15, K=NULL, platform= c("Visium", "ST", 'scRNAseq'),
       nfeatures=2000,K_set = seq(2, 10), variable.type=c("HVGs", "SVGs") ...)

Arguments

seu
an object of class "Seurat". The details of this object are given under 'Details'.

q
a positive integer, specify the number of latent features to be extracted, default as 15.

K
a positive integer, specify the number of clusters, default as NULL. When K=NULL, it is automatically selected by MBIC criteria.

platform
a string, specify the platform of the provided data, default as "Visium". There are three platforms to be chosen, including "Visuim", "ST" and "scRNAseq", where the first two means there are spatial coordinates information in the metadata of seu, named "row" and "col" and a Hidden Markov random field is used to model the spatial coordinates, and the last one "scRNAseq" means there is no spatial information in object seu and a multinomial model is used to model the unobserved class labels. The platform helps to calculate the adjoint matrix.

nfeatures
a positive integer, means how many highly variable or spatially variable genes used for DR-SC model, default as 2000. If there are less than 2000 features in seu, then all features are used for DR-SC model.

K_set
an optional vector of positive integer, means the candidates of number of clusters used for MBIC.

variable.type
an optional string, specify whether use highly variable genes (HVGs) or spatially variable genes (SVGs) to fit DR-SC model, default as "HVGs".

...
Other arguments to pass into DR.SC.fit function.

Details

seu is an object named Seurat, thich can easily created by R package Seurat. If the data is collected by the spatial transcriptomics technologies such as 10X Visium, ST, seqFISH, MERFISH and Slide-seq, then there are spatial coordinates information in the metadata of seu, named "row" and "col". DR-SC model uses a Hidden Markov random field to model the spatial coordinates. If the data is collected by the single cell RNA sequencing techonologies which means there is no spatial information in object seu then a multinomial model is used to model the unobserved class labels.

Value

DR.SC returns a revised Seurat object. There are two revisions in the seu. 1. the metadata is added a new column named spatial.drsc.cluster that represents the clustering results from DR-SC model. 2. a DimReduc object named dr-sc is added in the slot reductions, which represents the features extracted by DR-SC model.

References

Author

Wei Liu

Note

nothing

See also

Examples

## we generate the spatial transcriptomics data with lattice neighborhood, i.e. ST platform.
seu <- gendata_RNAExp(height=30, width=30,p=500, K=4,platform="ST")

#> Warning: The following arguments are not used: row.names

library(Seurat)
seu <- NormalizeData(seu)
# choose 400 spatially variable features
seu <- FindSVGs(seu, nfeatures = 400)

#> Loading required package: SPARK

#> Find the spatially variables genes by SPARK-X...
#> ## ===== SPARK-X INPUT INFORMATION ==== 
#> ## number of total samples: 900 
#> ## number of total genes: 500 
#> ## Running with single core, may take some time 
#> ## Testing With Projection Kernel
#> ## Testing With Gaussian Kernel 1
#> ## Testing With Gaussian Kernel 2
#> ## Testing With Gaussian Kernel 3
#> ## Testing With Gaussian Kernel 4
#> ## Testing With Gaussian Kernel 5
#> ## Testing With Cosine Kernel 1
#> ## Testing With Cosine Kernel 2
#> ## Testing With Cosine Kernel 3
#> ## Testing With Cosine Kernel 4
#> ## Testing With Cosine Kernel 5

seu1 <- DR.SC(seu, K=4, platform = 'ST', verbose=F)

#> Neighbors were identified for 900 out of 900 spots.

#> Warning: nrow(seu) is less than nfeatures, so assign nfeatures with nrow(seu)!

#> Warning: number of variable genes is less than nfeatures, 
#>               so assign nfeatures with number of variable genes!

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 't': 'NA' indices are not (yet?) supported for sparse Matrices

# choose spatially variable features
seu <- FindSVGs(seu)

#> Find the spatially variables genes by SPARK-X...
#> ## ===== SPARK-X INPUT INFORMATION ==== 
#> ## number of total samples: 900 
#> ## number of total genes: 500 
#> ## Running with single core, may take some time 
#> ## Testing With Projection Kernel
#> ## Testing With Gaussian Kernel 1
#> ## Testing With Gaussian Kernel 2
#> ## Testing With Gaussian Kernel 3
#> ## Testing With Gaussian Kernel 4
#> ## Testing With Gaussian Kernel 5
#> ## Testing With Cosine Kernel 1
#> ## Testing With Cosine Kernel 2
#> ## Testing With Cosine Kernel 3
#> ## Testing With Cosine Kernel 4
#> ## Testing With Cosine Kernel 5

# use SVGs to fit DR.SC model
seu1 <- DR.SC(seu, K=4,platform = 'ST', variable.type = 'SVGs',verbose=F)

#> Neighbors were identified for 900 out of 900 spots.

#> Warning: nrow(seu) is less than nfeatures, so assign nfeatures with nrow(seu)!

#> Using SVGs to fit DR.SC model since variable.type=SVGs...
#> Fit DR-SC model...
#> elasped time is : 15.18 1.09 16.22 NA NA 
#> Finish DR-SC model fitting

## genrate the scRNAseq data
seu <- gendata_RNAExp(height=30, width=30,p=500, K=4,platform="scRNAseq")

#> Warning: The following arguments are not used: row.names

seu <- NormalizeData(seu)
# choose 400 variable features using Seurat
seu <- FindVariableFeatures(seu, nfeatures = 400)
seu1 <- DR.SC(seu, K=4, platform = 'scRNAseq',nfeatures=400, verbose=F)

#> Fit DR-SC model...
#> elasped time is : 10.3 0.82 11.25 NA NA 
#> Finish DR-SC model fitting